# aurora-go/docs/swagger.yaml
basePath: /api
definitions:
...
  github_com_aurora-go_aurora_internal_vo.CommentVO:
    properties:
      aboutId:
        description: 关于页ID（type=3时使用）
        type: integer
      articleId:
        description: 文章ID（type=1时使用）
        type: integer
      commentContent:
        description: 评论内容（兼容前端 commentContent 字段名）
        type: string
      current:
        description: 当前页（对标Java MyBatis-Plus分页）
        type: integer
      friendLinkId:
        description: 友链ID（type=4时使用）
        type: integer
      parentId:
        description: 回复的父评论ID(0=顶级)
        type: integer
      replyUserId:
        description: 被回复用户ID
        type: integer
      size:
        description: 每页大小
        type: integer
      talkId:
        description: 说说ID（type=5时使用）
        type: integer
      topicId:
        description: 主题ID（前端传字符串，对标Java topicId）
        type: string
      type:
        enum:
        - 1
        - 2
        - 3
        - 4
        - 5
        type: integer
    required:
    - type
    type: object
...
host: localhost:8080
info:
  contact:
    email: aurora@example.com
    name: Aurora Team
  description: Aurora 博客系统 Go 1.26 后端 - 基于 tRPC-Agent-Go 的 AI 驱动博客平台
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://github.com/nicepkg/aurora
  title: Aurora Blog API
  version: "1.0"
paths:
...
  /api/admin/comments:
    delete:
      consumes:
      - application/json
      description: 后台批量删除评论
      parameters:
      - description: 评论ID列表
        in: body
        name: ids
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            type: object
        "400":
          description: 请求参数错误
          schema:
            type: object
        "401":
          description: 未授权/Token无效
          schema:
            type: object
        "500":
          description: 服务器内部错误
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 删除评论
      tags:
      - 评论
    get:
      consumes:
      - application/json
      description: 后台评论列表（含审核状态筛选）
      parameters:
      - description: 搜索关键词
        in: query
        name: keyword
        type: string
      - description: 当前页码
        in: query
        name: current
        type: integer
      - description: 每页数量
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回评论列表
          schema:
            type: object
        "401":
          description: 未授权/Token无效
          schema:
            type: object
        "500":
          description: 服务器内部错误
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 后台评论列表
      tags:
      - 评论
  /api/admin/comments/review:
    put:
      consumes:
      - application/json
      description: 批量审核评论（通过/拒绝）
      parameters:
      - description: 审核参数(isReview:审核状态, ids:评论ID列表)
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            type: object
        "400":
          description: 请求参数错误
          schema:
            type: object
        "401":
          description: 未授权/Token无效
          schema:
            type: object
        "500":
          description: 服务器内部错误
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 审核评论
      tags:
      - 评论
  /api/admin/comments/stats:
    get:
      consumes:
      - application/json
      description: 获取后台评论统计信息
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回评论统计数据
          schema:
            type: object
        "401":
          description: 未授权/Token无效
          schema:
            type: object
        "500":
          description: 服务器内部错误
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 获取评论统计信息
      tags:
      - 评论
  /api/admin/exception/logs:
    delete:
      consumes:
      - application/json
      description: 批量删除异常日志
      parameters:
      - description: 异常日志ID列表
        in: body
        name: ids
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            type: object
        "400":
          description: 请求参数错误
          schema:
            type: object
        "401":
          description: 未授权/Token无效
          schema:
            type: object
        "500":
          description: 服务器内部错误
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 批量删除异常日志
      tags:
      - 异常日志
    get:
      consumes:
      - application/json
      description: 获取异常日志列表（分页）
      parameters:
      - description: 搜索关键词
        in: query
        name: keyword
        type: string
      - description: 当前页码
        in: query
        name: current
        type: integer
      - description: 每页数量
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回异常日志列表
          schema:
            type: object
        "401":
          description: 未授权/Token无效
          schema:
            type: object
        "500":
          description: 服务器内部错误
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 获取异常日志列表
      tags:
      - 异常日志
...
  /api/comments:
    get:
      consumes:
      - application/json
      description: 获取前台评论列表（分页）
      parameters:
      - description: 评论类型
        in: query
        name: type
        type: integer
      - description: 主题ID
        in: query
        name: topicId
        type: integer
      - description: 当前页码
        in: query
        name: current
        type: integer
      - description: 每页数量
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回评论列表
          schema:
            type: object
        "500":
          description: 服务器内部错误
          schema:
            type: object
      summary: 获取评论列表
      tags:
      - 评论
  /api/comments/{id}/like:
    post:
      consumes:
      - application/json
      description: 点赞指定评论
      parameters:
      - description: 评论ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            type: object
        "400":
          description: 请求参数错误
          schema:
            type: object
        "401":
          description: 未授权/Token无效
          schema:
            type: object
        "500":
          description: 服务器内部错误
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 点赞评论
      tags:
      - 评论
  /api/comments/{id}/replies:
    get:
      consumes:
      - application/json
      description: 根据评论ID获取回复列表
      parameters:
      - description: 评论ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回回复列表
          schema:
            type: object
        "400":
          description: 请求参数错误
          schema:
            type: object
        "500":
          description: 服务器内部错误
          schema:
            type: object
      summary: 获取评论回复列表
      tags:
      - 评论
  /api/comments/{id}/reply:
    post:
      consumes:
      - application/json
      description: 回复指定评论（需要登录）
      parameters:
      - description: 父评论ID
        in: path
        name: id
        required: true
        type: integer
      - description: 回复内容
        in: body
        name: replyVO
        required: true
        schema:
          $ref: '#/definitions/github_com_aurora-go_aurora_internal_vo.CommentVO'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            type: object
        "400":
          description: 请求参数错误
          schema:
            type: object
        "401":
          description: 未授权/Token无效
          schema:
            type: object
        "500":
          description: 服务器内部错误
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 回复指定评论
      tags:
      - 评论
  /api/comments/save:
    post:
      consumes:
      - application/json
      description: 发表评论或回复评论（需要登录）
      parameters:
      - description: 评论内容
        in: body
        name: commentVO
        required: true
        schema:
          $ref: '#/definitions/github_com_aurora-go_aurora_internal_vo.CommentVO'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            type: object
        "400":
          description: 请求参数错误
          schema:
            type: object
        "401":
          description: 未授权/Token无效
          schema:
            type: object
        "500":
          description: 服务器内部错误
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 发表评论/回复
      tags:
      - 评论
  /api/comments/topSix:
    get:
      consumes:
      - application/json
      description: 获取前6条最新评论（用于侧边栏）
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回评论列表
          schema:
            type: object
        "500":
          description: 服务器内部错误
          schema:
            type: object
      summary: 获取前6条最新评论
      tags:
      - 评论
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Token 认证，格式: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"